Move token tests to own file
